switched SteamCommunityTest for GutoMaiaSteamCommunityTest
